backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime-gpu==1.18.0
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
